per #344, started breaking apart metplus.util.met_util to move functions into more appropriate files to better organize the METplus utilities
