+++ internal/tests/pytests/util/config_metplus/test_config_metplus.py
@@ -150,9 +150,9 @@ def test_find_var_indices_fcst(metplus_config,
     data_types = ['FCST']
     config.set('config', config_var_name, "NAME1")
     met_tool = 'grid_stat' if set_met_tool else None
-    var_name_indices = config_metplus.find_var_name_indices(config,
-                                                            data_types=data_types,
-                                                            met_tool=met_tool)
+    var_name_indices = config_metplus._find_var_name_indices(config,
+                                                             data_types=data_types,
+                                                             met_tool=met_tool)
 
     assert len(var_name_indices) == len(expected_indices)
     for actual_index in var_name_indices:
@@ -648,7 +648,7 @@ def test_find_var_indices_wrapper_specific(metplus_config, met_tool, indices):
     conf.set('config', f'{data_type}_VAR1_NAME', "NAME1")
     conf.set('config', f'{data_type}_GRID_STAT_VAR2_NAME', "GSNAME2")
 
-    var_name_indices = config_metplus.find_var_name_indices(conf, data_types=[data_type],
+    var_name_indices = config_metplus._find_var_name_indices(conf,data_types=[data_type],
                                                   met_tool=met_tool)
 
     assert var_name_indices == indices
@@ -1073,3 +1073,33 @@ def test_getraw_instance_with_unset_var(metplus_config):
     )
     new_config.set('config', 'CURRENT_FCST_NAME', 'NAME')
     assert new_config.getraw('config', 'OUTPUT_PREFIX') == 'FCST_NAME'
+
+
+@pytest.mark.parametrize(
+    'config_value, expected_result', [
+        # 2 items semi-colon at end
+        ('GRIB_lvl_typ = 234;  desc = "HI_CLOUD";',
+         'GRIB_lvl_typ = 234; desc = "HI_CLOUD";'),
+        # 2 items no semi-colon at end
+        ('GRIB_lvl_typ = 234;  desc = "HI_CLOUD"',
+         'GRIB_lvl_typ = 234; desc = "HI_CLOUD";'),
+        # 1 item semi-colon at end
+        ('GRIB_lvl_typ = 234;',
+         'GRIB_lvl_typ = 234;'),
+        # 1 item no semi-colon at end
+        ('GRIB_lvl_typ = 234',
+         'GRIB_lvl_typ = 234;'),
+    ]
+)
+@pytest.mark.util
+def test_format_var_items_options_semicolon(config_value,
+                                            expected_result):
+    time_info = {}
+
+    field_configs = {'name': 'FNAME',
+                     'levels': 'FLEVEL',
+                     'options': config_value}
+
+    var_items = config_metplus._format_var_items(field_configs, time_info)
+    result = var_items.get('extra')
+    assert result == expected_result
