+++ metplus/util/__init__.py
@@ -1,11 +1,13 @@
+from .metplus_check import *
 from .constants import *
 from .string_manip import *
-from .metplus_check import *
+from .system_util import *
+from .time_util import *
+from .string_template_substitution import *
 from .doc_util import *
 from .config_metplus import *
-from .time_util import *
+from .run_util import *
 from .met_util import *
-from .string_template_substitution import *
 from .met_config import *
 from .time_looping import *
 from .field_util import *
