+++ metplus/util/met_util.py
@@ -7,17 +7,16 @@ import gzip
 import bz2
 import zipfile
 import struct
-import getpass
+
 from dateutil.relativedelta import relativedelta
-from pathlib import Path
-from importlib import import_module
 
 from .string_manip import getlist, getlistint
+from .system_util import mkdir_p
+from . import time_util as time_util
 from .string_template_substitution import do_string_sub
 from .string_template_substitution import parse_template
-from . import time_util as time_util
 from .time_looping import time_generator
-from .. import get_metplus_version
+
 
 """!@namespace met_util
  @brief Provides  Utility functions for METplus.
@@ -25,398 +24,6 @@ from .. import get_metplus_version
 
 from .constants import *
 
-# missing data value used to check if integer values are not set
-# we often check for None if a variable is not set, but 0 and None
-# have the same result in a test. 0 may be a valid integer value
-MISSING_DATA_VALUE = -9999
-
-def pre_run_setup(config_inputs):
-    from . import config_metplus
-    version_number = get_metplus_version()
-    print(f'Starting METplus v{version_number}')
-
-    # Read config inputs and return a config instance
-    config = config_metplus.setup(config_inputs)
-
-    logger = config.logger
-
-    user_info = get_user_info()
-    user_string = f' as user {user_info} ' if user_info else ' '
-
-    config.set('config', 'METPLUS_VERSION', version_number)
-    logger.info('Running METplus v%s%swith command: %s',
-                version_number, user_string, ' '.join(sys.argv))
-
-    logger.info(f"Log file: {config.getstr('config', 'LOG_METPLUS')}")
-    logger.info(f"METplus Base: {config.getdir('METPLUS_BASE')}")
-    logger.info(f"Final Conf: {config.getstr('config', 'METPLUS_CONF')}")
-    config_list = config.getstr('config', 'CONFIG_INPUT').split(',')
-    for config_item in config_list:
-        logger.info(f"Config Input: {config_item}")
-
-    # validate configuration variables
-    isOK_A, isOK_B, isOK_C, isOK_D, all_sed_cmds = config_metplus.validate_configuration_variables(config)
-    if not (isOK_A and isOK_B and isOK_C and isOK_D):
-        # if any sed commands were generated, write them to the sed file
-        if all_sed_cmds:
-            sed_file = os.path.join(config.getdir('OUTPUT_BASE'), 'sed_commands.txt')
-            # remove if sed file exists
-            if os.path.exists(sed_file):
-                os.remove(sed_file)
-
-            write_list_to_file(sed_file, all_sed_cmds)
-            config.logger.error(f"Find/Replace commands have been generated in {sed_file}")
-
-        logger.error("Correct configuration variables and rerun. Exiting.")
-        sys.exit(1)
-
-    if not config.getdir('MET_INSTALL_DIR', must_exist=True):
-        logger.error('MET_INSTALL_DIR must be set correctly to run METplus')
-        sys.exit(1)
-
-    # set staging dir to OUTPUT_BASE/stage if not set
-    if not config.has_option('config', 'STAGING_DIR'):
-        config.set('config', 'STAGING_DIR',
-                   os.path.join(config.getdir('OUTPUT_BASE'), "stage"))
-
-    # handle dir to write temporary files
-    handle_tmp_dir(config)
-
-    # handle OMP_NUM_THREADS environment variable
-    handle_env_var_config(config,
-                          env_var_name='OMP_NUM_THREADS',
-                          config_name='OMP_NUM_THREADS')
-
-    config.env = os.environ.copy()
-
-    return config
-
-def run_metplus(config, process_list):
-    total_errors = 0
-
-    try:
-        processes = []
-        for process, instance in process_list:
-            try:
-                logname = f"{process}.{instance}" if instance else process
-                logger = config.log(logname)
-                package_name = ('metplus.wrappers.'
-                                f'{camel_to_underscore(process)}_wrapper')
-                module = import_module(package_name)
-                command_builder = (
-                    getattr(module, f"{process}Wrapper")(config,
-                                                         instance=instance)
-                )
-
-                # if Usage specified in PROCESS_LIST, print usage and exit
-                if process == 'Usage':
-                    command_builder.run_all_times()
-                    return 0
-            except AttributeError:
-                logger.error("There was a problem loading "
-                             f"{process} wrapper.")
-                return 1
-            except ModuleNotFoundError:
-                logger.error(f"Could not load {process} wrapper. "
-                             "Wrapper may have been disabled.")
-                return 1
-
-            processes.append(command_builder)
-
-        # check if all processes initialized correctly
-        allOK = True
-        for process in processes:
-            if not process.isOK:
-                allOK = False
-                class_name = process.__class__.__name__.replace('Wrapper', '')
-                logger.error("{} was not initialized properly".format(class_name))
-
-        # exit if any wrappers did not initialized properly
-        if not allOK:
-            logger.info("Refer to ERROR messages above to resolve issues.")
-            return 1
-
-        all_commands = []
-        for process in processes:
-            new_commands = process.run_all_times()
-            if new_commands:
-                all_commands.extend(new_commands)
-
-        # if process list contains any wrapper that should run commands
-        if any([item[0] not in NO_COMMAND_WRAPPERS for item in process_list]):
-            # write out all commands and environment variables to file
-            if not write_all_commands(all_commands, config):
-                # report an error if no commands were generated
-                total_errors += 1
-
-        # compute total number of errors that occurred and output results
-        for process in processes:
-            if process.errors != 0:
-                process_name = process.__class__.__name__.replace('Wrapper', '')
-                error_msg = '{} had {} error'.format(process_name, process.errors)
-                if process.errors > 1:
-                    error_msg += 's'
-                error_msg += '.'
-                logger.error(error_msg)
-                total_errors += process.errors
-
-        return total_errors
-    except:
-        logger.exception("Fatal error occurred")
-        logger.info(f"Check the log file for more information: {config.getstr('config', 'LOG_METPLUS')}")
-        return 1
-
-def post_run_cleanup(config, app_name, total_errors):
-    logger = config.logger
-    # scrub staging directory if requested
-    if (config.getbool('config', 'SCRUB_STAGING_DIR') and
-            os.path.exists(config.getdir('STAGING_DIR'))):
-        staging_dir = config.getdir('STAGING_DIR')
-        logger.info("Scrubbing staging dir: %s", staging_dir)
-        logger.info('Set SCRUB_STAGING_DIR to False to preserve '
-                    'intermediate files.')
-        shutil.rmtree(staging_dir)
-
-    # save log file path and clock time before writing final conf file
-    log_message = (f"Check the log file for more information: "
-                   f"{config.getstr('config', 'LOG_METPLUS')}")
-
-    start_clock_time = datetime.strptime(config.getstr('config', 'CLOCK_TIME'),
-                                         '%Y%m%d%H%M%S')
-
-    # rewrite final conf so it contains all of the default values used
-    write_final_conf(config)
-
-    # compute time it took to run
-    end_clock_time = datetime.now()
-    total_run_time = end_clock_time - start_clock_time
-    logger.debug(f"{app_name} took {total_run_time} to run.")
-
-    user_info = get_user_info()
-    user_string = f' as user {user_info}' if user_info else ''
-    if not total_errors:
-        logger.info(log_message)
-        logger.info('%s has successfully finished running%s.',
-                    app_name, user_string)
-        return
-
-    error_msg = (f'{app_name} has finished running{user_string} '
-                 f'but had {total_errors} error')
-    if total_errors > 1:
-        error_msg += 's'
-    error_msg += '.'
-    logger.error(error_msg)
-    logger.info(log_message)
-    sys.exit(1)
-
-def get_user_info():
-    """! Get user information from OS. Note that some OS cannot obtain user ID
-    and some cannot obtain username.
-    @returns username(uid) if both username and user ID can be read,
-     username if only username can be read, uid if only user ID can be read,
-     or an empty string if neither can be read.
-    """
-    try:
-        username = getpass.getuser()
-    except OSError:
-        username = None
-
-    try:
-        uid = os.getuid()
-    except AttributeError:
-        uid = None
-
-    if username and uid:
-        return f'{username}({uid})'
-
-    if username:
-        return username
-
-    if uid:
-        return uid
-
-    return ''
-
-def write_all_commands(all_commands, config):
-    """! Write all commands that were run to a file in the log
-     directory. This includes the environment variables that
-     were set before each command.
-
-    @param all_commands list of tuples with command run and
-     list of environment variables that were set
-    @param config METplusConfig object used to write log output
-     and get the log timestamp to name the output file
-    @returns False if no commands were provided, True otherwise
-    """
-    if not all_commands:
-        config.logger.error("No commands were run. "
-                            "Skip writing all_commands file")
-        return False
-
-    log_timestamp = config.getstr('config', 'LOG_TIMESTAMP')
-    filename = os.path.join(config.getdir('LOG_DIR'),
-                            f'.all_commands.{log_timestamp}')
-    config.logger.debug(f"Writing all commands and environment to {filename}")
-    with open(filename, 'w') as file_handle:
-        for command, envs in all_commands:
-            for env in envs:
-                file_handle.write(f"{env}\n")
-
-            file_handle.write("COMMAND:\n")
-            file_handle.write(f"{command}\n\n")
-
-    return True
-
-def handle_tmp_dir(config):
-    """! if env var MET_TMP_DIR is set, override config TMP_DIR with value
-     if it differs from what is set
-     get config temp dir using getdir_nocheck to bypass check for /path/to
-     this is done so the user can set env MET_TMP_DIR instead of config TMP_DIR
-     and config TMP_DIR will be set automatically"""
-    handle_env_var_config(config, 'MET_TMP_DIR', 'TMP_DIR')
-
-    # create temp dir if it doesn't exist already
-    # this will fail if TMP_DIR is not set correctly and
-    # env MET_TMP_DIR was not set
-    mkdir_p(config.getdir('TMP_DIR'))
-
-def handle_env_var_config(config, env_var_name, config_name):
-    """! If environment variable is set, use that value
-     for the config variable and warn if the previous config value differs
-
-     @param config METplusConfig object to read
-     @param env_var_name name of environment variable to read
-     @param config_name name of METplus config variable to check
-    """
-    env_var_value = os.environ.get(env_var_name, '')
-    config_value = config.getdir_nocheck(config_name, '')
-
-    # do nothing if environment variable is not set
-    if not env_var_value:
-        return
-
-    # override config config variable to environment variable value
-    config.set('config', config_name, env_var_value)
-
-    # if config config value differed from environment variable value, warn
-    if config_value != env_var_value:
-        config.logger.warning(f'Config variable {config_name} ({config_value}) '
-                              'will be overridden by the environment variable '
-                              f'{env_var_name} ({env_var_value})')
-
-def get_skip_times(config, wrapper_name=None):
-    """! Read SKIP_TIMES config variable and populate dictionary of times that should be skipped.
-         SKIP_TIMES should be in the format: "%m:begin_end_incr(3,11,1)", "%d:30,31", "%Y%m%d:20201031"
-         where each item inside quotes is a datetime format, colon, then a list of times in that format
-         to skip.
-         Args:
-             @param config configuration object to pull SKIP_TIMES
-             @param wrapper_name name of wrapper if supporting
-               skipping times only for certain wrappers, i.e. grid_stat
-             @returns dictionary containing times to skip
-    """
-    skip_times_dict = {}
-    skip_times_string = None
-
-    # if wrapper name is set, look for wrapper-specific _SKIP_TIMES variable
-    if wrapper_name:
-        skip_times_string = config.getstr('config',
-                                          f'{wrapper_name.upper()}_SKIP_TIMES', '')
-
-    # if skip times string has not been found, check for generic SKIP_TIMES
-    if not skip_times_string:
-        skip_times_string = config.getstr('config', 'SKIP_TIMES', '')
-
-        # if no generic SKIP_TIMES, return empty dictionary
-        if not skip_times_string:
-            return {}
-
-    # get list of skip items, but don't expand begin_end_incr yet
-    skip_list = getlist(skip_times_string, expand_begin_end_incr=False)
-
-    for skip_item in skip_list:
-        try:
-            time_format, skip_times = skip_item.split(':')
-
-            # get list of skip times for the time format, expanding begin_end_incr
-            skip_times_list = getlist(skip_times)
-
-            # if time format is already in skip times dictionary, extend list
-            if time_format in skip_times_dict:
-                skip_times_dict[time_format].extend(skip_times_list)
-            else:
-                skip_times_dict[time_format] = skip_times_list
-
-        except ValueError:
-            config.logger.error(f"SKIP_TIMES item does not match format: {skip_item}")
-            return None
-
-    return skip_times_dict
-
-def skip_time(time_info, skip_times):
-    """!Used to check the valid time of the current run time against list of times to skip.
-        Args:
-            @param time_info dictionary with time information to check
-            @param skip_times dictionary of times to skip, i.e. {'%d': [31]} means skip 31st day
-            @returns True if run time should be skipped, False if not
-    """
-    if not skip_times:
-        return False
-
-    for time_format, skip_time_list in skip_times.items():
-        # extract time information from valid time based on skip time format
-        run_time_value = time_info.get('valid')
-        if not run_time_value:
-            return False
-
-        run_time_value = run_time_value.strftime(time_format)
-
-        # loop over times to skip for this format and check if it matches
-        for skip_time in skip_time_list:
-            if int(run_time_value) == int(skip_time):
-                return True
-
-    # if skip time never matches, return False
-    return False
-
-def write_final_conf(config):
-    """! Write final conf file including default values that were set during
-     run. Move variables that are specific to the user's run to the [runtime]
-     section to avoid issues such as overwriting existing log files.
-
-        @param config METplusConfig object to write to file
-     """
-    final_conf = config.getstr('config', 'METPLUS_CONF')
-
-    # remove variables that start with CURRENT
-    config.remove_current_vars()
-
-    # move runtime variables to [runtime] section
-    config.move_runtime_configs()
-
-    config.logger.info('Overwrite final conf here: %s' % (final_conf,))
-    with open(final_conf, 'wt') as conf_file:
-        config.write(conf_file)
-
-def is_loop_by_init(config):
-    """!Check config variables to determine if looping by valid or init time"""
-    if config.has_option('config', 'LOOP_BY'):
-        loop_by = config.getstr('config', 'LOOP_BY').lower()
-        if loop_by in ['init', 'retro']:
-            return True
-        elif loop_by in ['valid', 'realtime']:
-            return False
-
-    if config.has_option('config', 'LOOP_BY_INIT'):
-        return config.getbool('config', 'LOOP_BY_INIT')
-
-    msg = 'MUST SET LOOP_BY to VALID, INIT, RETRO, or REALTIME'
-    if config.logger is None:
-        print(msg)
-    else:
-        config.logger.error(msg)
-
-    return None
 
 def loop_over_times_and_call(config, processes, custom=None):
     """! Loop over all run times and call wrappers listed in config
@@ -702,18 +309,6 @@ def round_0p5(val):
 
     return round(val * 2) / 2
 
-def mkdir_p(path):
-    """!
-       From stackoverflow.com/questions/600268/mkdir-p-functionality-in-python
-       Creates the entire directory path if it doesn't exist (including any
-       required intermediate directories).
-       Args:
-           @param path : The full directory path to be created
-       Returns
-           None: Creates the full directory path if it doesn't exist,
-                 does nothing otherwise.
-    """
-    Path(path).mkdir(parents=True, exist_ok=True)
 
 def get_storms(filter_filename, id_only=False, sort_column='STORM_ID'):
     """! Get each storm as identified by a column in the input file.
@@ -823,16 +418,6 @@ def prune_empty(output_dir, logger):
                              "...removing")
                 os.rmdir(full_dir)
 
-def camel_to_underscore(camel):
-    """! Change camel case notation to underscore notation, i.e. GridStatWrapper to grid_stat_wrapper
-         Multiple capital letters are excluded, i.e. PCPCombineWrapper to pcp_combine_wrapper
-         Numerals are also skipped, i.e. ASCII2NCWrapper to ascii2nc_wrapper
-         Args:
-             @param camel string to convert
-             @returns string in underscore notation
-    """
-    s1 = re.sub(r'([^\d])([A-Z][a-z]+)', r'\1_\2', camel)
-    return re.sub(r'([a-z])([A-Z])', r'\1_\2', s1).lower()
 
 def shift_time_seconds(time_str, shift):
     """ Adjust time by shift seconds. Format is %Y%m%d%H%M%S
@@ -845,168 +430,6 @@ def shift_time_seconds(time_str, shift):
     return (datetime.strptime(time_str, "%Y%m%d%H%M%S") +
             timedelta(seconds=shift)).strftime("%Y%m%d%H%M%S")
 
-def get_threshold_via_regex(thresh_string):
-    """!Ensure thresh values start with >,>=,==,!=,<,<=,gt,ge,eq,ne,lt,le and then a number
-        Optionally can have multiple comparison/number pairs separated with && or ||.
-        Args:
-            @param thresh_string: String to examine, i.e. <=3.4
-        Returns:
-            None if string does not match any valid comparison operators or does
-              not contain a number afterwards
-            regex match object with comparison operator in group 1 and
-            number in group 2 if valid
-    """
-
-    comparison_number_list = []
-    # split thresh string by || or &&
-    thresh_split = re.split(r'\|\||&&', thresh_string)
-    # check each threshold for validity
-    for thresh in thresh_split:
-        found_match = False
-        for comp in list(VALID_COMPARISONS)+list(VALID_COMPARISONS.values()):
-            # if valid, add to list of tuples
-            # must be one of the valid comparison operators followed by
-            # at least 1 digit or NA
-            if thresh == 'NA':
-                comparison_number_list.append((thresh, ''))
-                found_match = True
-                break
-
-            match = re.match(r'^('+comp+r')(.*\d.*)$', thresh)
-            if match:
-                comparison = match.group(1)
-                number = match.group(2)
-                # try to convert to float if it can, but allow string
-                try:
-                    number = float(number)
-                except ValueError:
-                    pass
-
-                comparison_number_list.append((comparison, number))
-                found_match = True
-                break
-
-        # if no match was found for the item, return None
-        if not found_match:
-            return None
-
-    if not comparison_number_list:
-        return None
-
-    return comparison_number_list
-
-
-def validate_thresholds(thresh_list):
-    """ Checks list of thresholds to ensure all of them have the correct format
-        Should be a comparison operator with number pair combined with || or &&
-        i.e. gt4 or >3&&<5 or gt3||lt1
-        Args:
-            @param thresh_list list of strings to check
-        Returns:
-            True if all items in the list are valid format, False if not
-    """
-    valid = True
-    for thresh in thresh_list:
-        match = get_threshold_via_regex(thresh)
-        if match is None:
-            valid = False
-
-    if valid is False:
-        print("ERROR: Threshold values must use >,>=,==,!=,<,<=,gt,ge,eq,ne,lt, or le with a number, "
-              "optionally combined with && or ||")
-        return False
-    return True
-
-def write_list_to_file(filename, output_list):
-    with open(filename, 'w+') as f:
-        for line in output_list:
-            f.write(f"{line}\n")
-
-def format_var_items(field_configs, time_info=None):
-    """! Substitute time information into field information and format values.
-
-        @param field_configs dictionary with config variable names to read
-        @param time_info dictionary containing time info for current run
-        @returns dictionary containing name, levels, and output_names, as
-         well as thresholds and extra options if found. If not enough
-         information was set in the METplusConfig object, an empty
-         dictionary is returned.
-    """
-    # dictionary to hold field (var) item info
-    var_items = {}
-
-    # set defaults for optional items
-    var_items['levels'] = []
-    var_items['thresh'] = []
-    var_items['extra'] = ''
-    var_items['output_names'] = []
-
-    # get name, return error string if not found
-    search_name = field_configs.get('name')
-    if not search_name:
-        return 'Name not found'
-
-    # perform string substitution on name
-    if time_info:
-        search_name = do_string_sub(search_name,
-                                    skip_missing_tags=True,
-                                    **time_info)
-    var_items['name'] = search_name
-
-    # get levels, performing string substitution on each item of list
-    for level in getlist(field_configs.get('levels')):
-        if time_info:
-            level = do_string_sub(level,
-                                  **time_info)
-        var_items['levels'].append(level)
-
-    # if no levels are found, add an empty string
-    if not var_items['levels']:
-        var_items['levels'].append('')
-
-    # get threshold list if it is set
-    # return error string if any thresholds not formatted properly
-    search_thresh = field_configs.get('thresh')
-    if search_thresh:
-        thresh = getlist(search_thresh)
-        if not validate_thresholds(thresh):
-            return 'Invalid threshold supplied'
-
-        var_items['thresh'] = thresh
-
-    # get extra options if it is set, format with semi-colons between items
-    search_extra = field_configs.get('options')
-    if search_extra:
-        if time_info:
-            search_extra = do_string_sub(search_extra,
-                                         **time_info)
-
-        # strip off empty space around each value
-        extra_list = [item.strip() for item in search_extra.split(';')]
-
-        # split up each item by semicolon, then add a semicolon to the end
-        # use list(filter(None to remove empty strings from list
-        extra_list = list(filter(None, extra_list))
-        var_items['extra'] = f"{'; '.join(extra_list)};"
-
-    # get output names if they are set
-    out_name_str = field_configs.get('output_names')
-
-    # use input name for each level if not set
-    if not out_name_str:
-        for _ in var_items['levels']:
-            var_items['output_names'].append(var_items['name'])
-    else:
-        for out_name in getlist(out_name_str):
-            if time_info:
-                out_name = do_string_sub(out_name,
-                                         **time_info)
-            var_items['output_names'].append(out_name)
-
-    if len(var_items['levels']) != len(var_items['output_names']):
-        return 'Number of levels does not match number of output names'
-
-    return var_items
 
 def sub_var_info(var_info, time_info):
     if not var_info:
@@ -1310,21 +733,6 @@ def template_to_regex(template, time_info):
     return do_string_sub(in_template,
                          **time_info)
 
-def is_python_script(name):
-    """ Check if field name is a python script by checking if any of the words
-     in the string end with .py
-
-     @param name string to check
-     @returns True if the name is determined to be a python script command
-     """
-    if not name:
-        return False
-
-    all_items = name.split(' ')
-    if any(item.endswith('.py') for item in all_items):
-        return True
-
-    return False
 
 def expand_int_string_to_list(int_string):
     """! Expand string into a list of integer values. Items are separated by
