+++ metplus/util/system_util.py
@@ -1,0 +1,58 @@
+"""
+Program Name: system_manip.py
+Contact(s): George McCabe
+Description: METplus utility to handle OS/system calls
+"""
+
+import os
+from pathlib import Path
+import getpass
+
+
+def mkdir_p(path):
+    """!
+       From stackoverflow.com/questions/600268/mkdir-p-functionality-in-python
+       Creates the entire directory path if it doesn't exist (including any
+       required intermediate directories).
+       Args:
+           @param path : The full directory path to be created
+       Returns
+           None: Creates the full directory path if it doesn't exist,
+                 does nothing otherwise.
+    """
+    Path(path).mkdir(parents=True, exist_ok=True)
+
+
+def get_user_info():
+    """! Get user information from OS. Note that some OS cannot obtain user ID
+    and some cannot obtain username.
+    @returns username(uid) if both username and user ID can be read,
+     username if only username can be read, uid if only user ID can be read,
+     or an empty string if neither can be read.
+    """
+    try:
+        username = getpass.getuser()
+    except OSError:
+        username = None
+
+    try:
+        uid = os.getuid()
+    except AttributeError:
+        uid = None
+
+    if username and uid:
+        return f'{username}({uid})'
+
+    if username:
+        return username
+
+    if uid:
+        return uid
+
+    return ''
+
+
+def write_list_to_file(filename, output_list):
+    with open(filename, 'w+') as f:
+        for line in output_list:
+            f.write(f"{line}\n")
