+++ metplus/wrappers/tc_gen_wrapper.py
@@ -16,7 +16,7 @@ import re
 
 from ..util import met_util as util
 from ..util import time_util
-from ..util import do_string_sub
+from ..util import do_string_sub, skip_time
 from ..util import time_generator
 from . import CommandBuilder
 
@@ -355,7 +355,7 @@ class TCGenWrapper(CommandBuilder):
             input_dict['custom'] = custom_string
             time_info = time_util.ti_calculate(input_dict)
 
-            if util.skip_time(time_info, self.c_dict.get('SKIP_TIMES', {})):
+            if skip_time(time_info, self.c_dict.get('SKIP_TIMES', {})):
                 self.logger.debug('Skipping run time')
                 continue
